Source
This is an executable Arduino code of the project. Here, we extract data from the accelerometer readings, then map these values (250-400) in sync with the range of a servo motor (0-180). The next step is to assign the mapped values to the servo. Lastly, we print all the acccelerometer readings  to better understand and validate the mechanism.